fix: Moved index.js from /user to /users
